04_03 Adds Input Dialog Service
1. Generate provider InputDialogService
2. Removes import for httpClient from input-dialog-service.ts as not needed, remove injection.
3. Adds Import InputDialogService to home.ts and injection
4. Adds Import GroceriesService to input-dialog-service.ts and injection
5. Adds Import AlertController to input-dialog-service.ts and injection
6. Extracts showAddItemPrompt and showEditItemPrompt from home.ts to input-dialog-service.ts
7. Adds updated call in editItem() and addItem() in home.ts to inputDialogService
8. Renames showEditItemPrompt() to showPrompt() in input-dialog-service.ts
9. Adds ? to make parameters optional in showPrompt() for add case.
10. Adds use of ternary null operator for showPrompt() function title, message, and inputs values
11. Adds if else loop for final action to if index !== undefined, then editItem(),
else addItem() in showPrompt() in input-dialog-service.ts
